requirements.txt
allure-pytest==2.15.0
allure-python-commons==2.15.0
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
colorama==0.4.6
greenlet==3.2.4
idna==3.11
iniconfig==2.3.0
packaging==25.0
playwright==1.55.0
pluggy==1.6.0
pyee==13.0.0
Pygments==2.19.2
pytest==8.4.2
requests==2.32.5
typing_extensions==4.15.0
urllib3==2.5.0
